+++ src/main/java/com/gregtechceu/gtceu/api/cover/filter/TagFluidFilter.java
@@ -4,9 +4,11 @@ import com.gregtechceu.gtceu.api.mui.factory.GuiData;
 import com.gregtechceu.gtceu.api.mui.value.sync.PanelSyncManager;
 import com.gregtechceu.gtceu.client.mui.screen.ModularPanel;
 import com.gregtechceu.gtceu.client.mui.screen.UISettings;
+import com.gregtechceu.gtceu.common.data.GTItems;
 import com.gregtechceu.gtceu.utils.TagExprFilter;
 
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraftforge.fluids.FluidStack;
@@ -31,21 +33,26 @@ public class TagFluidFilter extends TagFilter<FluidStack, FluidFilter> implement
     private static TagFluidFilter loadFilter(CompoundTag tag, Consumer<FluidFilter> itemWriter) {
         var handler = new TagFluidFilter();
         handler.itemWriter = itemWriter;
-        handler.oreDictFilterExpression = tag.getString("oreDict");
+        handler.filterString = tag.getString("oreDict");
         handler.matchExpr = null;
         handler.cache.clear();
-        handler.matchExpr = TagExprFilter.parseExpression(handler.oreDictFilterExpression);
+        handler.matchExpr = TagExprFilter.parseExpression(handler.filterString);
         return handler;
     }
 
-    public void setOreDict(String oreDict) {
+    public void setFilterString(String oreDict) {
         cache.clear();
-        super.setOreDict(oreDict);
+        super.setFilterString(oreDict);
+    }
+
+    @Override
+    protected ItemStack getFilterItem() {
+        return GTItems.TAG_FLUID_FILTER.asStack();
     }
 
     @Override
     public boolean test(FluidStack fluidStack) {
-        if (oreDictFilterExpression.isEmpty()) return false;
+        if (filterString.isEmpty()) return false;
         if (cache.containsKey(fluidStack.getFluid())) return cache.getOrDefault(fluidStack.getFluid(), false);
         if (TagExprFilter.tagsMatch(matchExpr, fluidStack)) {
             cache.put(fluidStack.getFluid(), true);
@@ -63,10 +70,5 @@ public class TagFluidFilter extends TagFilter<FluidStack, FluidFilter> implement
     @Override
     public boolean supportsAmounts() {
         return false;
-    }
-
-    @Override
-    public ModularPanel getPanel(GuiData data, PanelSyncManager syncManager, UISettings settings) {
-        return null;
     }
 }
