+++ src/main/java/com/gregtechceu/gtceu/utils/TagExprFilter.java
@@ -12,6 +12,12 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 public class TagExprFilter {
+    private static final Pattern DOUBLE_WILDCARD = Pattern.compile("\\*{2,}");
+    private static final Pattern DOUBLE_AND = Pattern.compile("&{2,}");
+    private static final Pattern DOUBLE_OR = Pattern.compile("\\|{2,}");
+    private static final Pattern DOUBLE_NOT = Pattern.compile("!{2,}");
+    private static final Pattern DOUBLE_XOR = Pattern.compile("\\^{2,}");
+    private static final Pattern DOUBLE_SPACE = Pattern.compile(" {2,}");
 
     public static class TagExprParser {
 
@@ -291,5 +297,60 @@ public class TagExprFilter {
                 .collect(Collectors.toSet());
 
         return expr != null && expr.matches(tags);
+    }
+
+    public static String tagInputValidator(String input) {
+        // remove all operators that are double
+        input = DOUBLE_WILDCARD.matcher(input).replaceAll("*");
+        input = DOUBLE_AND.matcher(input).replaceAll("&");
+        input = DOUBLE_OR.matcher(input).replaceAll("|");
+        input = DOUBLE_NOT.matcher(input).replaceAll("!");
+        input = DOUBLE_XOR.matcher(input).replaceAll("^");
+        input = DOUBLE_SPACE.matcher(input).replaceAll(" ");
+        // move ( and ) so it doesn't create invalid expressions f.e. xxx (& yyy) => xxx & (yyy)
+        // append or prepend ( and ) if the amount is not equal
+        StringBuilder builder = new StringBuilder();
+        int unclosed = 0;
+        char last = ' ';
+        for (int i = 0; i < input.length(); i++) {
+            char c = input.charAt(i);
+            if (c == ' ') {
+                if (last != '(')
+                    builder.append(" ");
+                continue;
+            }
+            if (c == '(')
+                unclosed++;
+            else if (c == ')') {
+                unclosed--;
+                if (last == '&' || last == '|' || last == '^') {
+                    int l = builder.lastIndexOf(" " + last);
+                    int l2 = builder.lastIndexOf(String.valueOf(last));
+                    builder.insert(l == l2 - 1 ? l : l2, ")");
+                    continue;
+                }
+                if (i > 0 && builder.charAt(builder.length() - 1) == ' ') {
+                    builder.deleteCharAt(builder.length() - 1);
+                }
+            } else if ((c == '&' || c == '|' || c == '^') && last == '(') {
+                builder.deleteCharAt(builder.lastIndexOf("("));
+                builder.append(c).append(" (");
+                continue;
+            }
+
+            builder.append(c);
+            last = c;
+        }
+        if (unclosed > 0) {
+            builder.append(")".repeat(unclosed));
+        } else if (unclosed < 0) {
+            unclosed = -unclosed;
+            for (int i = 0; i < unclosed; i++) {
+                builder.insert(0, "(");
+            }
+        }
+        input = builder.toString();
+        input = input.replaceAll(" {2,}", " ");
+        return input;
     }
 }
